logoutActions [nfc]: Move `logout` here, from `accountActions`
Along with its helper logoutPlain.

Soon, we'll get rid of the needsInitialFetch system, and call
register-queue code directly from thunk actions, including two in
accountActions.js: loginSuccess and accountSwitch.

Since that register-queue code can end up calling `logout`, and
since that code is (and indeed naturally belongs) outside
accountActions.js, we'd end up with an import cycle if `logout`
stayed in accountActions.js.

So, prevent that import cycle from happening by moving `logout` to a
new file.
